move edit label modal to cluster components
