+++ frontend/src/routes/ClusterManagement/Clusters/ClusterDetails/ClusterDetails.tsx
@@ -22,7 +22,7 @@ import { ClusterDeployment } from '../../../../resources/cluster-deployment'
 import { ManagedClusterInfo } from '../../../../resources/managed-cluster-info'
 import { CertificateSigningRequest } from '../../../../resources/certificate-signing-requests'
 import { ErrorPage } from '../../../../components/ErrorPage'
-import { EditLabelsModal } from '../../../../components/EditLabelsModal'
+import { EditLabelsModal } from '../components/EditLabelsModal'
 import { ResourceError, ResourceErrorCode } from '../../../../lib/resource-request'
 import { DownloadConfigurationDropdown } from '../components/DownloadConfigurationDropdown'
 import { ClusterManagementAddOn } from '../../../../resources/cluster-management-add-on'
